requirements.txt
Django==5.1.2
django-filter==24.3
django-graphql-jwt==0.4.0
graphene-django==3.2.2